MCS1 controller: in SmarActMCS1Wrapper.get_position. We rather use the ctypes c_int32 rather than c_long, otherwise the controller would not accept negative values. Works on Ubuntu, but has to be tested with Windows.

+++ src/pymodaq_plugins_smaract/hardware/smaract/mcs1/MCS1_wrapper.py
@@ -132,7 +132,7 @@ class SmarActMCS1Wrapper(object):
         position.value:signed int
         """
 
-        position = ct.c_long()
+        position = ct.c_int32()
 
         status = bindings.SA_GetPosition_S(
             ct.c_ulong(self.controller_index),
